Moved login and register form to folders
