+++ client/src/components/forms/loginForm/LoginForm.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useForm } from 'react-hook-form';
 import Button from '@material-ui/core/Button';
-import TextInput from '../inputs/textInput/TextInput';
 import { makeStyles } from '@material-ui/core';
+import TextInput from '../../inputs/textInput/TextInput';
 
 interface LoginFormInputs {
     email: string;
